Ajustes Diversos
Modifica a organização dos arquivos e ajusta o controle de movimento de elementos com o mouse

+++ src/utils/world/mouse.hpp
@@ -13,7 +13,7 @@ struct MouseInfo {
 // definições de mouse
 struct Mouse {
     int x, y; // posição do mouse em tempo real
-    int velX, velY; // velocidade instantânea do mouse em tempo real
+    double velX, velY; // velocidade instantânea do mouse em tempo real
     
     MouseInfo right;
     MouseInfo left;
